docs: Remove unfriendly disclaimer message
This patch removes the extremely unfriendly/unwelcoming message about this project. I don't know why it's here and feel that it's the wrong direction we should be taking our OSS repositories.

I'm happy to take on the responsibility of answering questions, addressing feature requests and looking at bug reports if everyone else is "too busy" to do so.