feat: Add `Code` component
This patch takes the existing `Highlight` component from the "demo app" and makes it part of the public API (available as `Code`).

Background: I need a `<Code>` component for an app I'm working on, and we have _many_ ways of handling this. My hope is adding this to `cauldron-react` will prevent us from creating any more React code highlighters in our code bases.

Closes #98